% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{extract_IsoSeq_ge}
\alias{extract_IsoSeq_ge}
\title{Extract the sequence around splice sites based on genome.}
\usage{
extract_IsoSeq_ge(ASdata, genome)
}
\arguments{
\item{ASdata}{A data frame holds an alternative splicing database,
including intron retention (IR), exon skipping (ES), alternative
donor sites (AltD), and alternative acceptor sites (AltA).}

\item{genome}{A BSgenome or XStringSet object.}
}
\value{
A alternative splicing data frame with sequence information.
}
\description{
A function for extracting a set of subsequences surrounding splice sites
from a sequence container like a BSgenome or XStraingSet.
}
\examples{
##Load the rice genome (MSU)
library("BSgenome.Osativa.MSU.MSU7")

##Load the alternative splicing database
path <- system.file("extdata","sample_riceAS.txt",
                     package = "AStrap")
rice_ASdata <-read.table(path,sep="\\t",head = TRUE,
                         stringsAsFactors = FALSE)

##Extract the sequence around splice sites based on genome
AS.data <- extract_IsoSeq_ge(rice_ASdata,Osativa)
}
