% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{extract_IsoSeq_tr}
\alias{extract_IsoSeq_tr}
\title{Extract the sequence around splice sites based on transcript.}
\usage{
extract_IsoSeq_tr(ASdata, transcriptSeq)
}
\arguments{
\item{ASdata}{A data frame holds the coordinate of splice sites with as least
three columns: Sname, Sstart, Send. Sname is the name of subject in the pariwise
sequence alignment; Sstart is the start of subject in the pairwise sequence
alignment; Send is the end of subject in the pairwise sequence;}

\item{transcriptSeq}{An XStringSet object holds the transcript sequence.}
}
\value{
A alternative splicing data frame with sequence information.
}
\description{
A function for extracting a set of subsequences surrounding splice sites
from a sequence container like XStraingSet.
}
\examples{
##Loading pairwise alignment data
load(system.file("data","sample_Aligndata.Rdata",package = "AStrap"))
head(Aligndata)

##Loading transcript sequence
trSequence.path <- system.file("extdata","example_TRsequence.fasta",
                              package = "AStrap")
trSequence <-  readDNAStringSet(trSequence.path,format = "fasta")

##Extract sequence around splice site based on the transcript sequence
Aligndata <- extract_IsoSeq_tr(Aligndata,trSequence )
colnames(Aligndata)
head(Aligndata$Ddown10)

head(Aligndata$Aup10)
}
