% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{getFeature}
\alias{getFeature}
\title{Feature construction.}
\usage{
getFeature(ASdata)
}
\arguments{
\item{ASdata}{A data frame with nine columns: donorSeq, acceptorSeq,
seq, length, Ddown10, Ddown20, Aup10, Aup20, Aup30.
donorSeq is the sequence of of the [-2, +3] region of donor sites;
acceptorSeq is the sequence of the [-2, +3] region of acceptor site;
seq is the sequence of the splice region;
length is the length of the splice region;
Ddown10 is the sequence of the downstream 10bp region of donor site;
Ddown20 is the sequence of the downstream 20bp region of donor site;
Aup10 is the sequence of the upstream 10bp region of acceptor site;
Aup20 is the sequence of the upstream 20bp region of acceptor site;
Aup30 is the sequence of the upstream 30bp region of acceptor site;}
}
\value{
A list of features.
}
\description{
Extract features from the sequences around splice sites,
  A total of 511 features.
}
\examples{
##Example 1: based on pairwise sequence alignment and transcript
##sequences.
##Loading pairwise alignment data
load(system.file("data","sample_Aligndata.Rdata",package = "AStrap"))
head(Aligndata)

##Loading transcript sequence
trSequence.path <- system.file("extdata","example_TRsequence.fasta",
                              package = "AStrap")
trSequence <-  readDNAStringSet(trSequence.path,format = "fasta")

##Extract sequence around splice site based on the transcript sequence
Aligndata <- extract_IsoSeq_tr(Aligndata,trSequence )
colnames(Aligndata)
head(Aligndata$Ddown10)
head(Aligndata$Aup10)
##Loading the consensus matrix of the sequences of the [-2,+3]
##region of acceptor sites
load(system.file("data","example_PWM_acceptor.Rdata",
                 package = "AStrap"))
##Loading the consensus matrix of the sequences of the [-2,+3]
##region of donor sites
load(system.file("data","example_PWM_donor.Rdata",
                 package = "AStrap"))
##Construction the feature space
feature <- getFeature(Aligndata)
ncol(feature)

##Example2: based on the alternative splicing database and genome.
path <- system.file("extdata","sample_riceAS.txt",package = "AStrap")
rice_ASdata <-read.table(path,sep="\\t",head = TRUE,
                         stringsAsFactors = FALSE)
head(rice_ASdata)
##Loading geneome using the package of BSgenome
library("BSgenome.Osativa.MSU.MSU7")
rice_ASdata<- extract_IsoSeq_ge(rice_ASdata,Osativa)
}
