% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{plotAlign}
\alias{plotAlign}
\title{Network graph showing identities of pairwise alignments in a cluster.}
\usage{
plotAlign(alignment, cluster.id = c("2"))
}
\arguments{
\item{alignment}{A data frame contation the information of sequence alignments.}

\item{cluster.id}{An interger vector indicating the ID of cluster.}
}
\value{
This function returns a network graph object invisibly.
}
\description{
This function plot pairwise alignments of isoforms of the
  same cluster network graph. It shows the relationship between isoforms
  of the same cluster, where the identity of a pairwise alignment is
  denoted as a line with different thickness and color. The higher
  the identity, the thicker the line. Red line denotes the identity of 1.
}
\examples{
##Loading example data
load(system.file("data","example_cluster.Rdata",package = "AStrap"))
g <- plotAlign(cluster.align$alignment,cluster.id=c("7","5"))
plot(g)
}
