% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{plotAS}
\alias{plotAS}
\title{Plotting a AS event.}
\usage{
plotAS(ASevent, id = 1)
}
\arguments{
\item{ASevent}{A list of AS events.}

\item{id}{The ID of AS events.}
}
\value{
This function retures a picture of AS event.
}
\description{
Visualization of AS events of different AS types, including AltA, AltD,
  ES, IR.
}
\examples{
##Load example data
load(system.file("data","sample_ASresult.Rdata",package = "AStrap"))
##Visualization of AS events of different AS types
library(Gviz)
plotAS(result$ASevent, id = 1)
plotAS(result$ASevent, id = 7)
plotAS(result$ASevent, id = 13)
plotAS(result$ASevent, id = 21)
}
