% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{plotCluster}
\alias{plotCluster}
\title{Network graph showing the distribution of isoforms in a cluster.}
\usage{
plotCluster(cluster, cluster.id = c("2"))
}
\arguments{
\item{cluster}{A data frame contation the information of isoform clusters.}

\item{cluster.id}{An interger vector indicating the ID of cluster.}
}
\value{
This function returns a network graph object invisibly.
}
\description{
This function plots a clustering network form a data frames.
 It shows the distribution of isoforms in a cluster.
}
\examples{
##Loading example data
load(system.file("data","example_cluster.Rdata",package = "AStrap"))
g<- plotCluster(cluster.align$cluster,cluster.id=c("7","5"))
plot(g)
}
