% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{plotGain}
\alias{plotGain}
\title{Attribute Evaluators.}
\usage{
plotGain(trainset, num)
}
\arguments{
\item{trainset}{The training data set.}

\item{num}{The number of top features based on the entropy value can
be display in figure.}
}
\description{
Using function InfoGainAttributeEval from Rweka  to
 evaluate the worth of an attribute by measuring the information
 gain with respect to the class.
}
\examples{
##Loading example data
rice_model<- load(system.file("data","rice_model.Rdata",package = "AStrap"))
library(RWeka)
library(ggplot2)
plotGain(trainset,20)
}
