% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{plotROC}
\alias{plotROC}
\title{Plotting a ROC curve}
\usage{
plotROC(probalility, labels)
}
\arguments{
\item{labels}{A vector, matrix, list, or data frame containing
the true class labels. Must have the same dimensions as 'predictions}

\item{probality}{A vector, matrix, list, or data frame containing
the predictions}
}
\value{
A ROC curve.
}
\description{
Plotting a ROC curve based on ggplots.
}
\examples{
##Loading example data
load(system.file("data","example_svm.Rdata",package = "AStrap"))
proba <- as.data.frame(attr(example_svm$svmPre, "probabilities"))
testset <- example_svm$testSet
label <- testset$class
library(ggplot2)
plotROC(proba,label)
}
