% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{readCDHIT}
\alias{readCDHIT}
\title{Read the clusters from a file.}
\usage{
readCDHIT(cdhit.path)
}
\arguments{
\item{cdhit.path}{A string for the name of a text file of
list of clusters output by CD-HIT-EST.}
}
\value{
A data frame contation the information of isoform
  clusters with four columns: ClusterID, seqID, seqLen, seqNum.
  ClusterID is the name of the cluster; seqID is the name of the
  transcript; seqLen is the length of the transcript; seqNum is
  the number of transcripts contatined in this class.
}
\description{
Read a text file of list of clusters output
  by CD-HIT-EST and creates a data frame form it.
}
\examples{
##load the exapmle data
cdhit.path <- system.file("extdata","example_cdhitest.clstr",
                           package = "AStrap")
raw.cluster <- readCDHIT(cdhit.path)
}
