% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStrap.R
\name{trimmGMAP}
\alias{trimmGMAP}
\title{Alignment trimming}
\usage{
trimmGMAP(gmap.file)
}
\arguments{
\item{gmap.file}{A string for the name of a GFF3 file of pairwise
sequence alignments output by GMAP.}
}
\value{
A data frame holds trimmed sequence alignment.
}
\description{
This function is to treat redundant sequence alignment
  and remove the invalid sequence alignments.
}
\examples{
##Load example data
gmap.path <- system.file("extdata","example_gmap.gff3",package = "AStrap")
trimmed.ali <- trimmGMAP(gmap.path)
}
